Hoist Advantage's "when attack ends" registration to InPlayCard
Move the just-in-time "When Attack/Defense Ends" ability machinery
(state field, getTriggeredAbilities inclusion, register/unregister, and
an overridable buildWhenAttackOrDefenseEndsAbility hook) onto InPlayCard,
the common superclass of units and upgrades. Advantage now only overrides
the build hook.

Merge the game-level attack-end listener into UnitPropertiesCard's
existing registerRulesListeners so there is a single rules-listener
registration, and drop Advantage's separate static registration in Game.

Remove the redundant isBlank() guard in the listener: gating the ability
behind the blank check in getTriggeredAbilities means a blanked token's
ability is never collected into the trigger window (TriggeredAbility
eventHandler checks getTriggeredAbilities().includes(this)).
